requirements.txt
sqlmodel
fastapi[standard]
pandas
scikit-learn
pillow